requirements.txt
fastapi
uvicorn[standard]
jinja2
asyncpg
ormar
psycopg2-binary
passlib
python-jose
python-multipart
typing
pydantic
jwt
httplib2
oauth2client
google-api-python-client
itsdangerous
httpx
